ActiScore Complete/AI Research Paper Summarizer/requirements.txt
flask
numpy
scikit-learn
nltk
reportlab
python-dotenv
PyPDF2
pdfplumber
python-magic